Only delete old order if it exists.

+++ registration/views/onsite_admin.py
@@ -537,8 +537,9 @@ def combine_orders(orders):
         for order_item in order_items:
             old_order = order_item.order
             order_item.order = first_order
-            logger.warning("Deleting old order id={0}".format(old_order.id))
-            old_order.delete()
+            if old_order:
+                logger.warning("Deleting old order id={0}".format(old_order.id))
+                old_order.delete()
             order_item.save()
 
         first_order.save()
